Infrastructure/weather-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg>=0.28.0
pyyaml==6.0.1
pydantic==2.5.0
httpx>=0.24.0
metar>=1.11.0












